feat(pkg/sqlite): add Open function, remove unused table name
