+++ cmd/gpud/command/utils.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/leptonai/gpud/components/state"
 	"github.com/leptonai/gpud/config"
+	"github.com/leptonai/gpud/pkg/sqlite"
 )
 
 func GetUID(ctx context.Context) (string, error) {
@@ -13,7 +14,7 @@ func GetUID(ctx context.Context) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("failed to get state file: %w", err)
 	}
-	db, err := state.Open(stateFile)
+	db, err := sqlite.Open(stateFile)
 	if err != nil {
 		return "", fmt.Errorf("failed to open state file: %w", err)
 	}
