+++ pkg/sqlite/sqlite.go
@@ -1,0 +1,22 @@
+// Package sqlite provides a SQLite3 database utils.
+package sqlite
+
+import (
+	"database/sql"
+	"fmt"
+
+	_ "github.com/mattn/go-sqlite3"
+)
+
+func Open(file string) (*sql.DB, error) {
+	// no need to run separate PRAGMA commands
+	// ref. https://www.sqlite.org/pragma.html#pragma_busy_timeout
+	// ref. https://www.sqlite.org/pragma.html#pragma_journal_mode
+	// ref. https://www.sqlite.org/pragma.html#pragma_synchronous
+	conns := fmt.Sprintf("%s?_busy_timeout=5000&_journal_mode=WAL&_synchronous=NORMAL", file)
+	db, err := sql.Open("sqlite3", conns)
+	if err != nil {
+		return nil, err
+	}
+	return db, nil
+}
